+++ doc/release_notes.rst
@@ -14,6 +14,8 @@ Upcoming Release
 
 * For industry distribution, use EPRTR as fallback if ETS data is not available.
 
+* New feature multi-decade optimisation with  perfect foresight.
+
 PyPSA-Eur 0.8.1 (27th July 2023)
 ================================